+++ qubes/tests/__init__.py
@@ -1442,6 +1442,7 @@ def load_tests(loader, tests, pattern):  # pylint: disable=unused-argument
             'qubes.tests.integ.dom0_update',
             'qubes.tests.integ.vm_update',
             'qubes.tests.integ.network',
+            'qubes.tests.integ.network_ipv6',
             'qubes.tests.integ.dispvm',
             'qubes.tests.integ.vm_qrexec_gui',
             'qubes.tests.integ.mime',
